# docs/spec/04_M1_OpenAPI.yaml
openapi: 3.1.0
info:
  title: 같이놀개 M1 API
  version: 1.3.1-m1
  description: |
    최신 제품 정책과 M1 통합 ERD를 기준으로 작성한 정적 REST 계약 정본이다.
    런타임 /v3/api-docs는 현재 구현 관찰용이며 계약 정본이나 필수 CI Gate가 아니다.
    M1은 폴링 방식이며 WebSocket, Push, 읽음 표시, 메시지 수정·삭제,
    사용자 셋로그 업로드, 회원탈퇴, 욕설 자동 차단 및 AI 대화 검열을 제공하지 않는다.
    최상위 bearerAuth가 적용되는 모든 operation은 인증 실패 시
    401 ErrorEnvelope를 반환한다.
    별도 HTTP Status는 이 정적 계약을 먼저 변경하고 ErrorCode·구현·테스트를 같은
    PR에서 동기화한다. 소유권 실패는 기본 403이며 존재 은닉이 필요한 endpoint만
    명시적으로 404를 사용한다.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /friend-requests:
    post:
      tags: [Friend]
      summary: 친구 요청
      operationId: createFriendRequest
      description: |
        Active Pet이 요청한다. 같은 방향의 유효한 PENDING은 409로 거절한다.
        반대 방향의 유효한 PENDING이 있으면 새 요청을 만들지 않고 기존 요청을
        ACCEPTED로 전이하고 Friendship과 DIRECT 방을 보장한다.
        만료 시각은 요청 시각부터 7일이며 expiresAt <= now이면 만료로 처리한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetPetId]
              properties:
                targetPetId:
                  type: integer
                  format: int64
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
        '201':
          description: PENDING 요청 생성. respondedAt과 directRoomId는 null이다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendRequestEnvelope'
        '200':
          description: |
            반대 방향의 기존 요청을 ACCEPTED로 전이하고 DIRECT 방을 보장한다.
            기존 requestId와 requester/target 방향을 유지하며 respondedAt과
            directRoomId를 반환한다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendRequestEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /friend-requests/received:
    get:
      tags: [Friend]
      summary: 받은 친구 요청 목록
      operationId: listReceivedFriendRequests
      description: |
        현재 Active Pet이 받은 유효한 PENDING 요청만 반환한다.
        expiresAt <= now인 요청은 목록에서 제외하지만 상태를 변경하지 않는다.
        requestedAt DESC, requestId DESC cursor를 사용한다.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 받은 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendRequestListEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /friend-requests/sent:
    get:
      tags: [Friend]
      summary: 보낸 친구 요청 목록
      operationId: listSentFriendRequests
      description: |
        현재 Active Pet이 보낸 유효한 PENDING 요청만 반환한다.
        expiresAt <= now인 요청은 목록에서 제외하지만 상태를 변경하지 않는다.
        requestedAt DESC, requestId DESC cursor를 사용한다.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 보낸 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendRequestListEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /friend-requests/{requestId}/accept:
    post:
      tags: [Friend]
      summary: 친구 요청 수락
      operationId: acceptFriendRequest
      description: |
        현재 Active Pet이 수신 Pet인 PENDING 요청만 수락한다. 요청자·수신자
        상태와 양방향 Block을 재검증하고 Friendship과 DIRECT 방을 같은
        트랜잭션에서 생성 또는 재사용한다. expiresAt <= now이면 요청을
        EXPIRED로 반영한 뒤 409를 반환한다. 권한이 없는 requestId는 404로
        존재를 숨긴다.
      parameters:
        - $ref: '#/components/parameters/FriendRequestId'
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 수락, Friendship 생성 및 DIRECT 방 보장
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendRequestEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /friend-requests/{requestId}/reject:
    post:
      tags: [Friend]
      summary: 친구 요청 거절
      operationId: rejectFriendRequest
      description: |
        현재 Active Pet이 수신 Pet인 PENDING 요청을 REJECTED로 전이한다.
        관계 종료 작업이므로 Block·친구 수·Friendship·Chat을 조회하지 않는다.
        expiresAt <= now이면 EXPIRED로 반영한 뒤 409를 반환한다. 권한이 없는
        requestId는 404로 존재를 숨긴다.
      parameters:
        - $ref: '#/components/parameters/FriendRequestId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 거절 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendRequestEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /friend-requests/{requestId}:
    delete:
      tags: [Friend]
      summary: 보낸 친구 요청 취소
      operationId: cancelFriendRequest
      description: |
        현재 Active Pet이 발신 Pet인 PENDING 요청을 CANCELED로 전이한다.
        관계 종료 작업이므로 Block·친구 수·Friendship·Chat·Pet 표시 정보를
        조회하지 않는다. expiresAt <= now이면 EXPIRED로 반영한 뒤 409를
        반환한다. 권한이 없는 requestId는 404로 존재를 숨긴다.
      parameters:
        - $ref: '#/components/parameters/FriendRequestId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '204':
          description: 취소 완료
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /pets/{petId}/friends/{friendPetId}:
    delete:
      tags: [Friend]
      summary: 친구 삭제
      operationId: deletePetFriend
      description: |
        인증 User가 소유한 미삭제 source Pet의 Friendship만 삭제한다.
        source Pet은 현재 Active Pet일 필요가 없으며 ACTIVE 또는 SUSPENDED를 허용한다.
        source Pet이 없거나 삭제됐으면 PET_NOT_FOUND, 미삭제 타인 소유 Pet이면
        PET_NOT_OWNED, canonical pair의 Friendship이 없으면 FRIENDSHIP_NOT_FOUND다.
        target Pet은 선행 조회하지 않는다. 기존 DIRECT ChatRoom, Participant,
        Message, Greeting, FriendRequest 이력은 변경하지 않으며 기존 Chat 조건을
        충족하는 참가자는 계속 대화할 수 있다.
      parameters:
        - $ref: '#/components/parameters/PetId'
        - name: friendPetId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '204':
          description: 친구 삭제 완료. 응답 Body 없음
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /chat/rooms/{roomId}/messages:
    get:
      tags: [Chat]
      summary: 새 메시지 폴링
      operationId: listChatMessages
      description: afterMessageId보다 큰 메시지를 ID 오름차순으로 반환한다.
      parameters:
        - $ref: '#/components/parameters/RoomId'
        - name: afterMessageId
          in: query
          required: false
          schema:
            type: integer
            format: int64
            minimum: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 메시지 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageListEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Chat]
      summary: TEXT 메시지 전송
      operationId: sendChatMessage
      description: |
        clientMessageId로 재시도 멱등성을 보장한다.
        ARCHIVED 방은 정상 메시지 저장 시 ACTIVE로 복구한다.
        대표 오류는 VALIDATION_FAILED(400), BLOCKED_USER(403),
        CHAT_ROOM_NOT_FOUND(404), GREETING_REPLY_REQUIRED·CHAT_DUPLICATE_MESSAGE(409)다.
      parameters:
        - $ref: '#/components/parameters/RoomId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageCreateRequest'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
        '201':
          description: 메시지 생성
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageEnvelope'
        '200':
          description: 동일 clientMessageId의 기존 메시지 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /me/blocks:
    post:
      tags: [Block]
      summary: User 단위 차단
      operationId: createBlock
      description: Friendship 삭제, 양방향 PENDING 요청 취소, 상호작용과 노출을 차단한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockCreateRequest'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '201':
          description: 차단 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockEnvelope'
        '200':
          description: 기존 차단 관계 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      tags: [Block]
      summary: 차단 목록
      operationId: listBlocks
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 차단 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 보호 API의 인증 실패는 401 ErrorEnvelope로 반환한다.
  parameters:
    PetId:
      name: petId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    FriendRequestId:
      name: requestId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: 권한 또는 정책 위반
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: 상태 또는 중복 충돌
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [success, message, data, error]
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
        data:
          type: 'null'
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    PetSearchItem:
      type: object
      required: [petId, publicTag, nickname, verified, relationship]
      properties:
        petId:
          type: integer
          format: int64
        publicTag:
          type: string
          maxLength: 30
          pattern: '^.{1,25}#[A-Z0-9]{4}$'
          description: Pet 공개 태그
        nickname:
          type: string
          minLength: 1
        profileUrl:
          type: [string, 'null']
          format: uri
          description: M1은 null이며 클라이언트가 기본 이미지를 표시한다.
        verified:
          type: boolean
        relationship:
          type: [string, 'null']
          enum: [NONE, REQUEST_SENT, REQUEST_RECEIVED, FRIEND, null]
          description: Active Pet이 없는 L1은 null
...
    FriendRequest:
      type: object
      required:
        - requestId
        - requesterPet
        - targetPet
        - status
        - requestedAt
        - respondedAt
        - expiresAt
        - directRoomId
      properties:
        requestId:
          type: integer
          format: int64
        requesterPet:
          $ref: '#/components/schemas/PetSearchItem'
        targetPet:
          $ref: '#/components/schemas/PetSearchItem'
        status:
          type: string
          enum: [PENDING, ACCEPTED, REJECTED, CANCELED, EXPIRED]
        requestedAt:
          type: string
          format: date-time
        respondedAt:
          type: [string, 'null']
          format: date-time
          description: ACCEPTED·REJECTED·CANCELED 전이 시각. PENDING·EXPIRED는 null
        expiresAt:
          type: string
          format: date-time
          description: 요청 시각부터 7일. expiresAt <= now이면 만료
        directRoomId:
          type: [integer, 'null']
          format: int64
          description: ACCEPTED 상태이면 생성 또는 재사용한 DIRECT 방 ID
...
    ChatMessageCreateRequest:
      type: object
      required: [clientMessageId, body]
      properties:
        clientMessageId:
          type: string
          maxLength: 64
        body:
          type: string
          minLength: 1
          maxLength: 2000
    ChatMessage:
      type: object
      required: [messageId, roomId, senderType, type, createdAt]
      properties:
        messageId:
          type: integer
          format: int64
        roomId:
          type: integer
          format: int64
        senderType:
          type: string
          enum: [PET, SYSTEM]
        senderPetId:
          type: [integer, 'null']
          format: int64
        type:
          type: string
          enum: [TEXT, CARD, SYSTEM]
        body:
          type: [string, 'null']
        meetingCardId:
          type: [integer, 'null']
          format: int64
        clientMessageId:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
...
    BlockCreateRequest:
      type: object
      required: [targetPetId]
      properties:
        targetPetId:
          type: integer
          format: int64
          description: 서버가 소유 User를 찾아 User 단위 차단으로 변환한다.
    Block:
      type: object
      required: [blockId, blockedUserId, createdAt]
      properties:
        blockId:
          type: integer
          format: int64
        blockedUserId:
          type: integer
          format: int64
        blockedUserPublicTag:
          type: string
        createdAt:
          type: string
          format: date-time
...
    CursorPage:
      type: object
      required: [nextCursor, hasNext]
      properties:
        nextCursor:
          type: [string, 'null']
        hasNext:
          type: boolean
...
    FriendRequestEnvelope:
      type: object
      required: [success, message, data, error]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data: { $ref: '#/components/schemas/FriendRequest' }
        error: { type: 'null' }
    FriendRequestListEnvelope:
      type: object
      required: [success, message, data, error]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data:
          type: object
          required: [items, page]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/FriendRequest' }
            page: { $ref: '#/components/schemas/CursorPage' }
        error: { type: 'null' }
...
    ChatMessageEnvelope:
      type: object
      required: [success, message, data, error]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data: { $ref: '#/components/schemas/ChatMessage' }
        error: { type: 'null' }
    ChatMessageListEnvelope:
      type: object
      required: [success, message, data, error]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data:
          type: object
          required: [items, nextAfterMessageId, hasMore]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/ChatMessage' }
            nextAfterMessageId:
              type: [integer, 'null']
              format: int64
            hasMore:
              type: boolean
        error: { type: 'null' }
...
    BlockEnvelope:
      type: object
      required: [success, message, data, error]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data: { $ref: '#/components/schemas/Block' }
        error: { type: 'null' }
    BlockListEnvelope:
      type: object
      required: [success, message, data, error]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data:
          type: object
          required: [items, page]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Block' }
            page: { $ref: '#/components/schemas/CursorPage' }
        error: { type: 'null' }
...